sdf: derive AssetPath serde, group Value conversions
Replace AssetPath's hand-written Serialize/Deserialize with a derived
#[serde(transparent)] over authored_path, marking the derived evaluated
and resolved paths #[serde(skip)] so they default to None on deserialize.

Move the Value <-> AssetPath conversions (TryFrom<Value> and From) out of
asset_path.rs into value.rs alongside the other Value conversions; the
module now holds only the type and its string-like and identity traits.
